requirements.txt
fastapi[standard]==0.112.0
pyyaml==6.0.1
pyinstaller==6.9.0
Pillow==10.4.0
pywebview==5.1
pystray==0.19.5
requests==2.32.3